moved temp utility functions to utils
